# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "unread",
            "required": false,
            "in": "query",
            "description": "true = เฉพาะที่ยังไม่อ่าน",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NotificationDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "การแจ้งเตือนของฉัน (ใหม่สุดก่อน)",
        "tags": [
          "notifications"
        ]
      },
      "patch": {
        "operationId": "NotificationsController_markAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/BulkUpdateResultDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ทำเครื่องหมายว่าอ่านแล้วทั้งหมด",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}": {
      "patch": {
        "operationId": "NotificationsController_mark",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/NotificationDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "อ่านแล้ว / ยังไม่อ่าน",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/v1/designs": {
      "get": {
        "operationId": "DesignsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "ค้นจากชื่องาน",
            "schema": {
              "example": "โปสเตอร์",
              "type": "string"
            }
          },
          {
            "name": "designType",
            "required": false,
            "in": "query",
            "schema": {
              "example": "poster",
              "type": "string"
            }
          },
          {
            "name": "folderId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "trashed",
            "required": false,
            "in": "query",
            "description": "true = งานในถังขยะ",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "editedWithin",
            "required": false,
            "in": "query",
            "description": "แก้ไขล่าสุดภายในช่วงนี้",
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "year"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "updated",
              "type": "string",
              "enum": [
                "updated",
                "created",
                "title"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DesignSummaryDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "งานของฉัน (ค้นหา กรองประเภท โฟลเดอร์ ช่วงวันที่แก้ หรือถังขยะ)",
        "tags": [
          "designs"
        ]
      },
      "post": {
        "operationId": "DesignsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDesignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DesignDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "สร้างงานใหม่ (หน้าเปล่า · จากเทมเพลต · หรือทำสำเนางานเดิม)",
        "tags": [
          "designs"
        ]
      }
    },
    "/api/v1/designs/{id}": {
      "get": {
        "operationId": "DesignsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DesignDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "งานหนึ่งชิ้นพร้อม JSON state (CMS ใช้ render สด)",
        "tags": [
          "designs"
        ]
      },
      "patch": {
        "operationId": "DesignsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDesignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DesignDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "บันทึกงาน · เปลี่ยนชื่อ · ย้ายโฟลเดอร์ · ย้ายลง/กู้คืนจากถังขยะ",
        "tags": [
          "designs"
        ]
      },
      "delete": {
        "operationId": "DesignsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบถาวร (เฉพาะงานที่อยู่ในถังขยะ)",
        "tags": [
          "designs"
        ]
      }
    },
    "/api/v1/design-type-usages": {
      "get": {
        "operationId": "DesignTypeUsagesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DesignTypeUsageDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ประเภทงานที่ฉันใช้บ่อย (นับจากงานจริง)",
        "tags": [
          "designs"
        ]
      }
    },
...
  },
  "info": {
    "title": "CS Canvas — csmju-canvas",
    "description": "ระบบย่อยภายใต้ CSMJU2030 · ตัวตนมาจาก JWT ของ Core Hub ที่ระบบนี้ตรวจลายเซ็นเองด้วย JWKS ทุก request (RS256 · iss=core-hub · aud=csmju2030 · kid=core-hub-2026) ส่ง token มาได้สองทาง: Authorization: Bearer หรือคุกกี้ csmju_canvas_access_token (ตั้งโดย /auth/callback)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csmju_canvas_access_token"
      }
    },
    "schemas": {
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "DESIGN_TRASHED",
              "TEMPLATE_USED"
            ]
          },
          "title": {
            "type": "string"
          },
          "link": {
            "type": "object",
            "nullable": true
          },
          "readAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "createdAt"
        ]
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 2480
          },
          "page": {
            "type": "number",
            "example": 2
          },
          "limit": {
            "type": "number",
            "example": 50
          },
          "totalPages": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "BulkUpdateResultDto": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "updated"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "read": {
            "type": "boolean",
            "description": "true = อ่านแล้ว · false = ทำเป็นยังไม่อ่าน"
          }
        },
        "required": [
          "read"
        ]
      },
      "DesignSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "designType": {
            "type": "string",
            "example": "poster"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "thumbnail": {
            "type": "object",
            "nullable": true
          },
          "folderId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "sourceTemplateId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "trashedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "designType",
          "width",
          "height",
          "createdAt",
          "updatedAt"
        ]
      },
      "DesignDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "designType": {
            "type": "string",
            "example": "poster"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "thumbnail": {
            "type": "object",
            "nullable": true
          },
          "folderId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "sourceTemplateId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "trashedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "document": {
            "type": "object",
            "description": "JSON state ของผืนผ้าใบ",
            "additionalProperties": true
          }
        },
        "required": [
          "id",
          "title",
          "designType",
          "width",
          "height",
          "createdAt",
          "updatedAt",
          "document"
        ]
      },
      "CreateDesignDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "โปสเตอร์งานสัปดาห์วิทยาศาสตร์"
          },
          "designType": {
            "type": "string",
            "example": "poster",
            "description": "ไม่ต้องส่งถ้าสร้างจากเทมเพลต"
          },
          "width": {
            "type": "number",
            "example": 1080
          },
          "height": {
            "type": "number",
            "example": 1350
          },
          "document": {
            "type": "object",
            "description": "JSON state ของผืนผ้าใบ · ไม่ส่ง = หน้าเปล่า"
          },
          "templateId": {
            "type": "string",
            "format": "uuid",
            "description": "สร้างเป็นสำเนาของเทมเพลตนี้"
          },
          "copyFromDesignId": {
            "type": "string",
            "format": "uuid",
            "description": "ทำสำเนาจากงานของตัวเอง"
          },
          "folderId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateDesignDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "document": {
            "type": "object",
            "description": "JSON state ของผืนผ้าใบทั้งก้อน"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "thumbnail": {
            "type": "string",
            "description": "data URL ของภาพย่อ"
          },
          "folderId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "null = เอาออกจากโฟลเดอร์"
          },
          "trashed": {
            "type": "boolean",
            "description": "true = ย้ายไปถังขยะ · false = กู้คืน"
          }
        }
      },
      "DeletedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
      "DesignTypeUsageDto": {
        "type": "object",
        "properties": {
          "designType": {
            "type": "string",
            "example": "presentation"
          },
          "count": {
            "type": "number",
            "example": 4
          },
          "lastUsedAt": {
            "type": "string"
          }
        },
        "required": [
          "designType",
          "count",
          "lastUsedAt"
        ]
      },
...
    }
  }
}